replica/table: increase the memory usage estimate for a view update
The current code estimates the amout of semaphore units needed
for a single view update to be 256 + (frozen mutation size in bytes).

It appears that this is not enough.
There's a test which causes Scylla to go OOM because of view updates,
even though each update waits for this amount of semaphore units.

Changing the 256 to 65536 fixes the problem - Scylla no longer
reports bad_allocs and doesn't crash. Why 65536? I tried 32768
and it still crashed, with 65536 the test passed without any problems.

The test went like this:
1) Start a 3-node cluster
2) Create keyspace with RF=1
3) Create a table
4) Create 50 materialized views based on this table
5) Create 50 partitions with 1000 rows each
6) Delete all the partitions in a single query (pk in (1, 2, ...))

This caused a multitude of view updates to be spawned at once,
and Scylla ran out of memory, which caused a crash.

The test was taken from this comment:
https://github.com/scylladb/scylladb/issues/12379#issuecomment-1378992818

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>