keys: change from_optional_exploded to accept a span instead of a vector
A span is more generic than a vector, and can be constructed
from any contiguous container (like small_vector), or a subset
of a container.

To support this, helpers in compound.hh need to use make_iterator_range,
since a span doesn't fit the container concept (since spans don't own
their contents).

This is needed to make a similar change to function evaluation, as
the token function passes its parameters to from_optional_exploded().