replica/table: don't drop view updates when there are no semaphore units
Due to recent changes, the code responsible for generating view updates
started waiting on the view_update_concurrency_semaphore.

The current logic, which this commit changes, reasons that if the
view_update_concurrency_semaphore has no units, then something went
horribly wrong and we should just panic and drop the view update.

I think this reasoning is flawed. There could be queries that eat up
all of view_update_concurrency_semaphore's units, while being fully
compliant with throttling etc.

Imagine a query which deletes a whole partition holding 10 million rows.
This query will need to generate 10 million view updates, one for each row.
Each of those updates will require consuming some units from the semaphore,
so the semaphore could run out of units because of a single, valid query.

With the current logic this scenario would cause us to drop all of the
view updates for queries which run in parallel to the ongoing DELETE.
Since the semaphore has no units even a single INSERT will decide
to skip the view update and cause a view inconsistency.

I think this is a bug. Running two valid queries in parallel shouldn't
result in view inconsistencies, no matter what those queries do.

To deal with this we could wait on the semaphore instead of skipping
the update when there are no units. This is what this commit does.

This change flips the current design on its head. There will be no more
dropped view updates.
I think that this is a better way to do materialized views. What good
are materialized views that sometimes decide to not recognize an update?
I think that as a user I would like be notified when my write was written
only to the base table. The views should be eventually consistent,
but there is no base-view repair, so in case of a dropped view update
ill the inconsistency ever be fixed?
I plan to put more arguments for this change in the github PR containing this commit.

Note that this doesn't really break high-availability, the remote updates are generated
and then spawned in the background. If a remote node dies then the update will be written
as a hint. The hint might be eventually deleted to free up space, but maybe read-repair
will repair the data based on another view replica.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>