replica/table: rename memory_usage_of to units_needed
The function memory_usage_of used to calculate an estimate
of how much memory is needed to perform the view updates.

Now the value has been significantly increased, and I think
it doesn't really represent the amount of memory needed.

The amount of units calculated by this function is simple
the amount of untis that's needed to avoid crashes on heavy
workloads. It doesn't necessary refer to memory, and was
chosen by trial and error so that tests don't crash.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>