doc: remove load-and-stream option from 5.1
Related: https://github.com/scylladb/scylla-enterprise/issues/2807

This commit removes the --load-and-stream nodetool option
from version 5.1 - it is not supported in this version.

This commit should only be merged to branch-5.1 (not to master)
as the feature will be added in the later versions => in versions
prior to 5.2.x the information about the option is a bug.