keys: print "non-utf8-key" when clustering_key is not UTF-8
before this change we do not check if the clustering_key to be formatted
is UTF-8 encoded before printing it. but we do perform the validation
when printing paritition_keys. since the clustering_key is not different
from partition_key when it comes to encoding, actually they are
different parts of a parimary key. so let's validate the encoding of
clustering_key as well, when formatting it. this change is a follow-up
of 85b21ba049d2a7439deec6a1133b006547145a24.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>