gms: Persist features explicitly in features enabler
Nowadays features are persisted in feature_service::enable() and there
are four callers of it

- feature enabler via gossiper notifications
- boot kicks feature enabler too
- schema loader tool
- cql test env

None but the first case need to persist features. The loader tool in
fact doesn't do it even now it by keeping qctx uninitialized. Cql test
env wires up the qctx, but it makes no differences for the test cases
themselves if the features are persisted or not.

Boot-time is a bit trickier -- it loads the feature list from system
keyspace and may filter-out some of them, then enable. In that case
committing the list back into system keyspace makes no sense, as the
persisted list doesn't extend.

The enabler, in turn, can call system keyspace directly via its explicit
dependency reference. This fixes the inverse dependency between system
keyspace and feature service.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>