cql3/stats: use zero-initialization
use {} instead of {0ul} for zero initialization. as `_query_cnt`
is a multi-dimension array, each elements in `_query_cnt` is yet
another array. so we cannot initialize it with a `{0ul}`. but
to zero-initialize this array, we can just use `{}`, as per
https://en.cppreference.com/w/cpp/language/zero_initialization

> If T is array type, each element is zero-initialized.

so this should recursively zero-initialize all arrays in `_query_cnt`.

this change should silence following warning:

stats.hh:88:60: error: suggest braces around initialization of subobject [-Werror,-Wmissing-braces]
            [statements::statement_type::MAX_VALUE + 1] = {0ul};
                                                           ^~~
                                                           {  }

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>